+++ sandbox/k8s/tf-sourceloop-sandbox/postgres.tf
@@ -4,9 +4,3 @@ resource "kubectl_manifest" "postgres_manifests" {
   yaml_body  = file(each.key)
   depends_on = [kubernetes_namespace.sourceloop_sandbox]
 }
-
-resource "kubectl_manifest" "pgadmin_manifests" {
-  for_each   = fileset(path.module, "pgadmin/*.yaml")
-  yaml_body  = file(each.key)
-  depends_on = [kubernetes_namespace.sourceloop_sandbox]
-}