+++ services/authentication-service/src/modules/auth/providers/keycloak-verify.provider.ts
@@ -66,7 +66,8 @@ export class KeycloakVerifyProvider
       if (
         !creds ||
         creds.authProvider !== 'keycloak' ||
-        creds.authId !== profile.keycloakId
+        (creds.authId !== profile.keycloakId &&
+          creds.authId !== profile.username)
       ) {
         throw new HttpErrors.Unauthorized(AuthErrorKeys.InvalidCredentials);
       }