+++ services/payment-service/src/models/index.ts
@@ -1,0 +1,4 @@
+export * from './transactions.model';
+export * from './orders.model';
+export * from './payment-gateways.model';
+export * from './templates.model';